+++ protocols/identify/CHANGELOG.md
@@ -5,9 +5,12 @@
 - Move I/O from `Behaviour` to `Handler`. Handle `Behaviour`'s Identify and Push requests independently by incoming order,
   previously Push requests were prioritized. see [PR 3208].
 
-- Update to `libp2p-swarm` `v0.42.0`.
+- Update to `libp2p-swarm` `v0.42.0`. Update to the `libp2p_swarm::handler::ConnectionEvent` `DialTimeout`
+  introduction and consequential changes. With that introduce `Event::Timeout`. See [PR 3307].
+
 
 [PR 3208]: https://github.com/libp2p/rust-libp2p/pull/3208
+[PR 3307]: https://github.com/libp2p/rust-libp2p/pull/3307
 
 # 0.41.1